+++ src/main/kotlin/com/github/surpsg/diffcoverage/domain/DiffCoverageConfiguration.kt
@@ -1,9 +1,9 @@
 package com.github.surpsg.diffcoverage.domain
 
 data class DiffCoverageConfiguration(
-    val project: String,
-    val projectDir: String,
-    val execFiles: List<String>,
-    val classes: List<String>,
-    val reportsRoot: String
+    val project: String = "",
+    val projectDir: String = "",
+    val execFiles: List<String> = listOf(),
+    val classes: List<String> = listOf(),
+    val reportsRoot: String = ""
 )