+++ src/main/kotlin/com/github/surpsg/diffcoverage/properties/BundleKeys.kt
@@ -1,10 +1,7 @@
 package com.github.surpsg.diffcoverage.properties
 
 const val PLUGIN_NAME = "name"
-const val REPORT_LINK = "diff.coverage.report.link"
-const val REPORT_LINK_WITH_ERROR = "diff.coverage.report.link.with.error"
 
 const val DIFF_COVERAGE_COLLECT_INFO = "collect.diff.coverage.info"
-const val DIFF_COVERAGE_COLLECT_INFO_FAILED = "collect.diff.coverage.info.failed"
 const val NOT_GRADLE_PROJECT = "not.gradle.project"
 const val NO_COVERAGE_FILES = "no.coverage.data.files"